+++ src/util.js
@@ -4,6 +4,7 @@ export const typeOf = (obj) =>
 export const isArr = (x) => typeOf(x) === "array";
 export const isEmpty = (x = []) => x.length <= 0;
 export const isFunc = (x) => typeof x === "function";
+export const isNil = (x) => [null, undefined].includes(x);
 export const isObj = (x) => typeOf(x) === "object";
 
 export const isSelector = (x) => isObj(x) || typeOf(x) === "string";
@@ -33,9 +34,20 @@ export function assign(prev, added) {
   );
 }
 
+/* Return all elements from first list that are not present in the second one. */
+export function difference(toKeep, toRemove) {
+  return toKeep.filter((item) => toRemove.indexOf(item) < 0);
+}
+
 /* Check whether or not a source array includes any of the searched arrray values. */
 export function includesSome(searchedArr, sourceArr) {
   return sourceArr.some((el) => searchedArr.indexOf(el) >= 0);
+}
+
+/* Combine elements from two lists, without duplicates. */
+export function union(arr1, arr2) {
+  const itemsToAdd = arr2.filter((item) => arr1.indexOf(item) < 0);
+  return [...arr1, ...itemsToAdd];
 }
 
 /* eslint-disable no-console */